// constants/Types.ts
// Definimos los tipos para los datos del formulario
export interface FormData {
    Make: string;
    Model: string;
    Year: string;
    Engine_Fuel_Type: string;
    Engine_HP: string;
    Engine_Cylinders: string;
    Transmission_Type: string;
    Driven_Wheels: string;
    Number_of_Doors: string;
    Vehicle_Size: string;
    Vehicle_Style: string;
    Highway_MPG: string;
    City_MPG: string;
    Popularity: string;
  }